pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rlwatch"
version = "0.1.0"
description = "Real-time GRPO/PPO training instability detection for ML teams"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "rlwatch", email = "hello@rlwatch.dev"}
]
keywords = ["reinforcement-learning", "GRPO", "PPO", "training", "monitoring", "debugging"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyyaml>=6.0",
    "scipy>=1.10.0",
    "rich>=13.0.0",
    "slack-sdk>=3.20.0",
    "streamlit>=1.30.0",
    "plotly>=5.18.0",
    "click>=8.1.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
torch = ["torch>=2.0.0"]
trl = ["trl>=0.7.0", "transformers>=4.35.0"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-benchmark>=4.0.0",
    "hypothesis>=6.90.0",
    "responses>=0.24.0",
]

[project.scripts]
rlwatch = "rlwatch.cli:main"

[project.urls]
Homepage = "https://github.com/rlwatch/rlwatch"
Documentation = "https://github.com/rlwatch/rlwatch#readme"
Issues = "https://github.com/rlwatch/rlwatch/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "perf: performance benchmark tests (Tier 5)",
    "integration: integration tests touching real SQLite/SMTP/HTTP (Tier 4)",
    "property: Hypothesis property-based tests (Tier 2)",
    "trl: requires the [trl] extra (transformers + trl)",
]
filterwarnings = [
    "ignore::DeprecationWarning:streamlit",
]

[tool.coverage.run]
source = ["src/rlwatch"]
omit = [
    # Streamlit dashboard is hard to test in-process; covered manually + by
    # the dashboard smoke test in CI.
    "*/dashboard.py",
]

[tool.coverage.report]
fail_under = 90
show_missing = true
skip_covered = false
